+++ examples/User/JsonApi/Resource/UserResource.php
@@ -128,7 +128,7 @@ class UserResource extends AbstractResource
                         ->setDataAsCallable(function () use ($user) {
                             return $user["contacts"];
                         }, $this->contactTransformer)
-                        ->omitWhenNotIncluded()
+                        ->omitDataWhenNotIncluded()
                     ;
             }
         ];